feat: Add support for clearing multiple slots in MySlot
This commit modifies the `options.lua` file to add support for clearing multiple slots in MySlot. Previously, only a single slot could be cleared at a time. With this change, the `MySlot:Clear` function now accepts an additional argument, which is a table containing the slots to be cleared. This allows for clearing multiple slots in a single operation.