feat: Add support for petslot in MySlot
This commit adds support for the `petslot` field in the `MySlot` class. The `petslot` field is a repeated field in the `Charactor` message of the `MySlot.proto` file. It allows for storing and retrieving pet action information in the `MySlot` addon. The `GetPetActionInfo` and `RecoverData` functions are modified to handle the `petslot` field.